Reorganized unit tests for messages and schemas pallets
