+++ pallets/messages/src/lib.rs
@@ -42,14 +42,10 @@
 	missing_docs
 )]
 
-#[cfg(test)]
-mod mock;
-
-#[cfg(test)]
-mod tests;
-
 #[cfg(feature = "runtime-benchmarks")]
 mod benchmarking;
+#[cfg(test)]
+mod tests;
 
 pub mod weights;
